Modify ZOrder system to different ZOrder siblings
This breaks Z-fighting guarantees and is therefore most likely
unsuitable for upstream.

The old system meant that sibling elements of an overlay container were
given a Z-ordering one more than the previously added sibling, with the
sibling to be added first given the Z-order one more than the container.
This meant that to control the Z-ordering of siblings one must strictly
control the order in which elements are added to the overlay container,
and once an element has been added it was difficult to change its
Z-order.

Now the meaning of the return value to _notifyZOrder has been changed to
be the Z-order of the callee, instead of the next available Z-order.
The Z-order of an element within a container is the Z-order of the
container plus a positive Z-order increment that is settable on a
per-element level, with 1 as the default. This means that the children
control their own Z-order within the container, unlike before.

Because the same Z-ordering process that applied to the children of an
overlay container applied to the children of an overlay, previously no
two elements could have the same Z-ordering and Z-fighting would not
occur. This change no longer guarantees the absence of Z-fighting, and
it is even possible for two siblings to have the same Z-order. Since
Z-order is used to determine render queue priority, and (as far as I
know) two renderables with the same priority have unspecified rendering
order, it is the responsibility of the user to avoid Z-fighting by
setting the Z-order increment of elements appropriately.