Support separate translation/scale/rotation tracks
Some file formats (e.g. nif) allow the animator to specify, for any given
keyframe, only a change to translation, or to scale, or to rotation. If
interpolation between keyframes is performed in a nonlinear fashion,
converting to a scheme where all three are required for every keyframe
is difficult. In particular, inserting the interpolated values of the
unspecified parts of the transform is insufficient; one must ensure that
the new keyframes (from the point of view of the original animation) are
inserted without disturbing the original interpolation. For a generic
interpolation scheme this may not even be possible (consider forcing the
interpolant to have a stationary point at each keyframe).

This commit provides a straightforward way around this problem by
simply adding new animation tracks which support changing just one of
translation, scale, or rotation. An alternative (and less intrusive to
OGRE) approach would be to use AnimableValues wrapping the parameters of
the animating node, but because the NumericAnimationTrack requires a
shared_ptr to the AnimableValue this meant several hoops to jump through
on the user's side. One would need to heap allocate a separate adaptor
for each node parameter, each of which having a pointer to the
particular value in the node. The user would have to manually construct
these for every entity before they could use any animation, and really
it seemed like a bit of a fuss.