Stop TerrainGroup from deleting ImportData
tldr: Makes all Terrain reloadable, even if it was created from data
in memory, at a memory cost or one-line inconvenience to users who
don't need Terrain reloading.

The documentation on TerrainGroup::unloadTerrain() warns that once
Terrain is unloaded it cannot be reloaded unless it was originally
loaded from a file, because the ImportData needed to recreate it is
destroyed once the Terrain is originally loaded. The only reason for
this appears to be to save space, but the user has no input on whether
that is even necessary.

If it *is* necessary, then the user can easily delete the ImportData
themselves; It is much harder for the user to work around the ImportData
deletion and implement their own reloading system than it is to call
TerrainGroup::getTerrainDefinition(...)->freeImportData().
I do not think it is unreasonable to expect that at least some users
(certainly at least one) would like reloadable manual Terrain.

ImportData is still deleted for data read from files.