Remove TagPoint dependency on Entity
While semantically a TagPoint should be attached to an Entity on one
side and a MovableObject on the other, the TagPoint itself doesn't
require that the Entity actually be one; a MovableObject is sufficient,
and the caller can be more precise if needed. This change allows
user-defined classes to make use of TagPoints, instead of their use
being reserved for Ogre::Entity.

Unfortunately a little bit of ugliness is needed, in that MovableObject
needs a virtual detachObjectFromBone method, even if generically it
doesn't make sense for a MovableObject to have bones.