Remove missing scheme/technique material fallback
The original behaviour was to fall back on the first technique when no
technique was available for the active material scheme. This is
reasonable behaviour if schemes are used for quality selection, but
wrong if the schemes are being used to select which objects to render,
such as when forward rendering transparent objects in a deferred renderer.
The deferred rendering sample circumvents this by giving every material
every scheme, leaving the undesired ones blank, but in programs where the
deferred materials are not being generated programmatically such work is
unnecessary.

This is a breaking change and is unlikely to be suitable for upstream
use.